+++ test/dummy/app/controllers/braintree/payment_methods_controller.rb
@@ -3,8 +3,8 @@ class Braintree::PaymentMethodsController < ApplicationController
   end
 
   def update
-    current_user.processor = params[:processor]
-    current_user.update_payment_method(params[:card_token])
+    current_user.set_payment_processor params[:processor]
+    current_user.payment_processor.update_payment_method(params[:card_token])
     redirect_to payment_method_path
   end
 end