+++ test/dummy/app/controllers/braintree/subscriptions_controller.rb
@@ -2,7 +2,7 @@ class Braintree::SubscriptionsController < ApplicationController
   before_action :set_subscription, only: [:show, :edit, :update, :destroy, :cancel, :resume]
 
   def index
-    @subscriptions = Pay::Subscription.where(processor: :braintree).order(created_at: :desc)
+    @subscriptions = Pay::Subscription.joins(:customer).where(pay_customers: {processor: :braintree}).order(created_at: :desc)
   end
 
   def show
@@ -12,9 +12,9 @@ class Braintree::SubscriptionsController < ApplicationController
   end
 
   def create
-    current_user.processor = params[:processor]
-    current_user.card_token = params[:card_token]
-    subscription = current_user.subscribe(plan: params[:plan_id])
+    current_user.set_payment_processor params[:processor]
+    current_user.payment_processor.payment_method_token = params[:card_token]
+    subscription = current_user.payment_processor.subscribe(plan: params[:plan_id])
     redirect_to braintree_subscription_path(subscription)
   rescue Pay::Error => e
     flash[:alert] = e.message
@@ -45,6 +45,6 @@ class Braintree::SubscriptionsController < ApplicationController
   private
 
   def set_subscription
-    @subscription = Pay::Subscription.where(processor: :braintree).find(params[:id])
+    @subscription = Pay::Subscription.find(params[:id])
   end
 end