+++ test/dummy/app/controllers/stripe/checkouts_controller.rb
@@ -1,8 +1,8 @@
 module Stripe
   class CheckoutsController < ApplicationController
     def show
-      current_user.processor = :stripe
-      current_user.customer
+      current_user.set_payment_processor :stripe
+      current_user.payment_processor.customer
 
       @payment = current_user.payment_processor.checkout(mode: "payment", line_items: "price_1ILVZaKXBGcbgpbZQ26kgXWG")
       @subscription = current_user.payment_processor.checkout(mode: "subscription", line_items: "default")