+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ### Unreleased
 
 * Update to `@hotwired/stimulus` for payments view - @excid3
+* Update test/dummy app to Pay 3 - @excid3
 
 ### 3.0.21